Data Standardization and Data Normalization
In this installment of Data Wrangling, we delve into the critical processes of data standardization and normalization. Standardizing and normalizing data enable better comparisons, eliminate scale-related biases, and improve the performance of machine learning models. Join us as we uncover the power of these techniques for a more robust data analysis journey! #DataWrangling #DataStandardization #DataNormalization #DataAnalysis"