Scatter Plot | Line Chart | Bar Plot | Bubble Plot
Plotly is a powerful and versatile open-source library used for creating interactive and visually appealing data visualizations in Python. It allows users to create a wide range of charts, graphs, and maps with ease. in this notebook, I covered these 8 plots
Scatter Plots
- Line Charts
- Bar Plot
- Bubble Plot
- Box Plot
- Histograms
- Distplots
- Heatmaps

Just a practice notebook, which I've learned from a youtube channel - part 1